requirements.txt
streamlit==1.34.0
openai==1.30.5
httpx<0.28
tiktoken
faiss-cpu==1.8.0
langchain==0.2.2
pandas
pdfplumber
PyMuPDF
pytesseract
pillow
tqdm
python-dotenv